moved extensions nearer to usage
